backend/app/requirements.txt
fastapi
uvicorn
httpx
python-dotenv
openai
pydantic
aiofiles
python-multipart
apscheduler
